audio_manager/pyproject.toml
[project]
name = "audio-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "pyodbc",
    "python-dotenv",
    "httpx",
    "pydantic",
    "boto3",
    "dependency-injector",
    "python-gitlab",
]

[project.scripts]
audio-manager = "audio_manager.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/audio_manager"]
